+++ src/last-key-of.ts
@@ -1,10 +1,22 @@
+import { invalidObject, notEnoughArgs } from "./errors";
 import hasOwn from "./has-own";
+import isObject from "./is-object";
 import { Key } from "./types";
 
 function lastKeyOf(object: {}, value: any): null;
 function lastKeyOf<K extends Key>(object: Record<K, any>, value: any): K | null;
 
 function lastKeyOf<K extends Key>(object: Record<K, any>, value: any): K | null {
+
+  const argsLen = arguments.length;
+
+  if (argsLen < 2) {
+    throw notEnoughArgs(argsLen, 2);
+  }
+
+  if (!isObject(object)) {
+    throw invalidObject(object);
+  }
 
   let result: K | null = null;
 
@@ -13,7 +25,7 @@ function lastKeyOf<K extends Key>(object: Record<K, any>, value: any): K | null 
       hasOwn.call(object, key) &&
       object[key] === value
     ) {
-      result = key;
+      result = key as K;
     }
   }