+++ src/map.ts
@@ -1,4 +1,6 @@
+import { invalidObject, notEnoughArgs } from "./errors";
 import hasOwn from "./has-own";
+import isObject from "./is-object";
 import { Extra, Key, MapCallback } from "./types";
 import { wrapFilterCallback } from "./wrap-callback";
 
@@ -21,6 +23,17 @@ function map<V, K extends Key, E extends Extra, RV = any, TH = any>(
   object: Record<K, V>,
   callback: MapCallback<V, K, E, TH, RV>,
 ): Record<K, RV> {
+
+  const args = arguments;
+  const argsLen = args.length;
+
+  if (argsLen < 2) {
+    throw notEnoughArgs(argsLen, 2);
+  }
+
+  if (!isObject(object)) {
+    throw invalidObject(object);
+  }
 
   const cb = wrapFilterCallback<V, K, E, TH, RV>(
     callback,