+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * improved documentation
 * performance improvements for none & 1 extra arguments
+* dependencies updated
 
 # 1.1.13 (01-06-2019)