+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-# 1.1.14 (04-06-2019)
+# 2.0.0 (05-06-2019)
 
 * improved documentation
 * performance improvements for none & 1 extra arguments
 * dependencies updated
 * `reduce` signatures improved
-* types refactoring
+* types major improvements
 * implemented error check
 
 # 1.1.13 (01-06-2019)